Change buttons to have only bottom borders
The contrast of the borders and backgrounds with the panel didn't feel right and this change removes the background color and provides only a bottom border (with padding to account for border loss). The bottom border is thicker on the focused/urgent workspaces and is not visible on workspaces with no windows. The urgent workspace has text and border are colored the same to draw attention.