Improve empty/urgent workspace handling
Doing extra if checks inside the set of if checks that applies styling seemed wasteful and meant applying the inactiveBtn style in multiple checks, so the check to determine empty workspaces is done beforehand. The whole 0=false/1=true thing also meant assigning strings to urgent/empty workspaces to be sure of checking against a specific string instead of true/false or 0/1.